Separate offset_encoding
